src/pyproject.toml
[project]
name = "dbms-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.27.1",
    "dotenv>=0.9.9",
    "elastic-transport>=9.2.0",
    "elasticsearch>=9.2.0",
    "ipykernel==6.29.5",
    "jupyter==1.1.1",
    "matplotlib>=3.10.7",
    "notebook==7.2.2",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pgvector>=0.4.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydicom>=3.0.1",
    "requests==2.32.3",
    "tqdm==4.67.1",
    "urllib3>=2.5.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=3.0.1",
    "open-clip-torch==2.23.0",
]

[project.optional-dependencies]
machine-learning = [
    "torch==2.9.0",
    "torchaudio==2.9.0",
    "torchvision==0.24.0",
    "transformers==4.35.2",
    "open-clip-torch==2.23.0",
]
